requirements.txt
oletools
PyPDF2
scikit-learn
pandas
numpy
requests
python-dotenv
pyzipper
reportlab